remove redundant structs
