+++ libs/protocol/src/node.rs
@@ -4,17 +4,26 @@
 use minotari_node_grpc_client::grpc::NodeIdentity;
 use serde::{Deserialize, Serialize};
 use tari_common_types::{emoji::EmojiId, types::PublicKey};
-use tari_comms::peer_manager::NodeId;
 use tari_utilities::byte_array::ByteArray;
+use tari_utilities::hex::Hex;
 
 use crate::wallet::InvalidPublicKey;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
 pub struct BaseNodeIdentity {
-    pub public_key: String,
-    pub public_addresses: String,
-    pub node_id: String,
+    pub public_key: Vec<u8>,
+    pub public_addresses: Vec<String>,
+    pub node_id: Vec<u8>,
     pub emoji_id: String,
+}
+
+impl BaseNodeIdentity {
+    /// Provide the base node connection string. It is of the form
+    /// "0eefb45a4de9484eca74846a4f47d2c8d38e76be1fec63b0112bd00d297c0928::/ip4/13.40.98.39/tcp/18189"
+    pub fn connection_string(&self) -> String {
+        format!("{}::/dns4/base_node/tcp/18189", self.public_key.to_hex())
+    }
 }
 
 impl TryFrom<NodeIdentity> for BaseNodeIdentity {
@@ -22,11 +31,12 @@ impl TryFrom<NodeIdentity> for BaseNodeIdentity {
 
     fn try_from(value: NodeIdentity) -> Result<Self, Self::Error> {
         let public_key = PublicKey::from_vec(&value.public_key).map_err(|e| InvalidPublicKey(e.to_string()))?;
+        // TODO: Implement `Serialize` for `EmojiId`
         let emoji_id = EmojiId::from_public_key(&public_key).to_string();
         Ok(BaseNodeIdentity {
-            public_key: public_key.to_string(),
-            public_addresses: value.public_addresses.join("::"),
-            node_id: NodeId::from_key(&public_key).to_string(),
+            public_key: value.public_key,
+            public_addresses: value.public_addresses,
+            node_id: value.node_id,
             emoji_id,
         })
     }
